data-processing/requirements.txt
google-generativeai
pillow
pdf2image
PyMuPDF
firebase-admin
python-dotenv